--- app/components/atoms/icons/icon.tsx
@@ -1,28 +1,0 @@
-// import React from 'react';
-// import { IconProps } from './types';
-
-// export const Diamond: React.FC<IconProps> = ({ size = 18, ...rest }) => {
-//   return (
-//     <svg width="40" height="40" viewBox="0 0 40 40" fill="none" xmlns="http://www.w3.org/2000/svg">
-//         <g clip-path="url(#clip0_8781_1154)">
-//         <path d="M36.2273 37H3.77273C3.34773 37 3 36.6357 3 36.1905V3.80952C3 3.36429 3.34773 3 3.77273 3H36.2273C36.6523 3 37 3.36429 37 3.80952V36.1905C37 36.6357 36.6523 37 36.2273 37ZM4.54545 35.381H35.4545V4.61905H4.54545V35.381Z" fill="#2233D0"/>
-//         <path d="M34.6815 7.85718H5.31787V9.47623H34.6815V7.85718Z" fill="#2233D0"/>
-//         <path d="M7.63627 5.42859H6.09082V7.04764H7.63627V5.42859Z" fill="#2233D0"/>
-//         <path d="M9.95463 5.42859H8.40918V7.04764H9.95463V5.42859Z" fill="#2233D0"/>
-//         <path d="M12.2725 5.42859H10.7271V7.04764H12.2725V5.42859Z" fill="#2233D0"/>
-//         <path d="M23.8637 32.1429H22.3183V24.0476H18.4546V32.1429H16.9092V22.4286H23.8637V32.1429Z" fill="#2233D0"/>
-//         <path d="M29.2724 32.1429H27.727V17.5714H23.8633V32.1429H22.3179V15.9524H29.2724V32.1429Z" fill="#2233D0"/>
-//         <path d="M18.4545 32.1429H16.9091V26.4762H13.0455V32.1429H11.5V24.8572H18.4545V32.1429Z" fill="#2233D0"/>
-//         <path d="M31.5912 31.3333H9.18213V32.9523H31.5912V31.3333Z" fill="#2233D0"/>
-//         <path d="M20.7729 18.3808H19.2274V14.3332H15.3638V12.7141H20.7729V18.3808Z" fill="#2233D0"/>
-//         <path d="M19.4519 12.9504L10.5674 22.2579L11.6602 23.4028L20.5447 14.0952L19.4519 12.9504Z" fill="#2233D0"/>
-//         </g>
-//         <defs>
-//         <clipPath id="clip0_8781_1154">
-//         <rect width="34" height="34" fill="white" transform="translate(3 3)"/>
-//         </clipPath>
-//         </defs>
-//     </svg>
-
-//   );
-// };
