--- app/components/molecules/button/button.module.css
@@ -1,28 +1,0 @@
-/* Button Default */
-.mth-btn-default{
-    background-color: #563CFE;
-    border: 1px solid #563CFE;
-    color: #fff;
-    filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.25));
-}
-
-.mth-btn-border{
-  background-color: #fff;
-  color: #563CFE;
-  border: 1px solid #563CFE;
-  filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.25));
-}
-
-.mth-btn{
-  padding: 10px;
-  cursor: pointer;
-}
-  
-.mth-btn-full{
-  width: 100%;
-}
-
-button:disabled{
-  background-color: #aba8b0;
-  border: 1px solid #aba8b0;
-}
