--- app/components/molecules/image/image.types.tsx
@@ -1,7 +1,0 @@
-export interface ImageProps {
-    src:string, 
-    alt:string, 
-    width?:number, 
-    height?:number, 
-     classIm?:string
-}
