--- app/components/molecules/input/passwordInput.module.css
@@ -1,9 +1,0 @@
-.mth-password{
-    position: absolute;
-    right: 10px;
-    top: 50%;
-    transform: translateY(-50%);
-    background: transparent;
-    border: none;
-    cursor: pointer;
-}
