--- app/components/molecules/listValidation/listValidationPassword.types.tsx
@@ -1,9 +1,0 @@
-export interface ValidationListProps {
-    validations: {
-      hasMinLength: boolean;
-      hasNumber: boolean;
-      hasUpperCase: boolean;
-      hasLowerCase: boolean;
-      hasSpecialChar: boolean;
-    };
-  }
