--- app/components/molecules/progressBar/progressBar.module.css
@@ -1,7 +1,0 @@
-.progress-check{
-  background-color: #0BFFC4;
-}
-
-.progress-no-check{
-  background-color: #0BFFFF;
-}
