--- app/components/molecules/select/optionSelect.tsx
@@ -1,5 +1,0 @@
-export interface optionSelect {
-    value: string;
-    label: string;
-    
-}
