--- app/components/molecules/select/selectForm.types.tsx
@@ -1,10 +1,0 @@
-export interface Option {
-  value: string;
-  label: string;
-}
-
-export interface SelectFormProps extends React.SelectHTMLAttributes<HTMLSelectElement> {
-  options: Option[];
-  label?: string;
-  error?: string;
-}
