--- app/components/organisms/card/card.module.css
@@ -1,9 +1,0 @@
-.cardFormat:hover {
-  cursor: pointer;
-  box-shadow: 0px 12px 30px 0px rgba(0, 0, 0, 0.2);
-  transition: all 800ms cubic-bezier(0.19, 1, 0.22, 1);
-}
-
-.cardFormat{
-  transition: all 500ms ease-in-out;
-}
