--- app/components/organisms/card/card.types.tsx
@@ -1,15 +1,0 @@
-export interface CardProps {
-    image: string, 
-    title: string,
-    description: string
-}
-
-export interface CardIndicatorProps {
-    logo: string, 
-    title: string,
-    number:number,
-    start:number,
-    end:number,
-    current:number
-
-}
