--- app/components/organisms/modal/modal.module.css
@@ -1,31 +1,0 @@
-/* modal */
-.overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.7);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.modal {
-  background: white;
-  padding: 25px;
-  border-radius: 5px;
-  width: auto; 
-  max-width: 65%; 
-  max-height: 82%;
-  /* overflow-y: auto;  */
-}
-
-.closeButton {
-  background: transparent;
-  color: #6f6565;
-  border: none;
-  font-size: 30px;
-  cursor: pointer;
-  float: right;
-}
