--- app/components/organisms/modal/modal.types.tsx
@@ -1,6 +1,0 @@
-export interface ModalProps {
-    isOpen: boolean;
-    onClose: () => void;
-    children: React.ReactNode;
-    width?: string;
-}
