--- app/components/organisms/table/tableOfContens.types.tsx
@@ -1,17 +1,0 @@
-import type { IconType } from "react-icons";
-
-// Actualiza la interfaz para incluir sublistas
-export interface MenuItem {
-    id: string;
-    title: string;
-    href: string;
-    isActive?: boolean;
-    icon?: IconType;
-    subItems?: MenuItem[]; // Agregamos subItems para las sublistas
-  }
-  
-  // Define las props del componente
-  export interface TableOfContentsProps {
-    items: MenuItem[];
-  }
-
