+++ app/main.tsx
@@ -2,7 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './app';
-import './app.css';
+// Estilos del proyecto
+import './output.css';
+// Estilos de la libreria
+import 'ui-mathilde-web/dist/style.css';
+
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
