--- app/styles/styleAtoms.css
@@ -1,34 +1,0 @@
-/* Mensajes alertas */
-
-  .mth-msg{
-    color: #380182;
-    margin-bottom: 1em;
-  }
-
-  /* Otros */
-
-  .justify-between{
-    justify-content: space-between;
-  }
-
-  .mth-line{
-    width: 48%;
-    border: 0.5px solid #563CFE;
-  }
-
-/* Input */
-
-.mth-input-code{
-  font-size: 32px;
-}
-
-input[type=number]::-webkit-inner-spin-button,
-input[type=number]::-webkit-outer-spin-button {
-    -webkit-appearance: none;
-    margin: 0;
-}
-
-input[type=number] {
-    -moz-appearance: textfield;
-    appearance: textfield;
-}
