+++ package.json
@@ -13,15 +13,12 @@
     "@originjs/vite-plugin-federation": "^1.3.9",
     "@react-router/node": "^7.1.3",
     "@react-router/serve": "^7.1.3",
-    "@tailwindcss/vite": "^4.0.0",
+    "@tailwindcss/vite": "^4.0.17",
     "@vitejs/plugin-react": "^4.3.4",
     "axios": "^1.7.9",
-    "chart.js": "^4.4.7",
-    "flowbite": "^3.1.2",
     "formik": "^2.4.6",
     "isbot": "^5.1.17",
     "react": "^18.3.1",
-    "react-chartjs-2": "^5.3.0",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.54.2",
     "react-icons": "^5.4.0",
@@ -29,7 +26,7 @@
     "react-router": "^7.1.3",
     "react-router-dom": "^7.1.3",
     "sweetalert2": "^11.15.10",
-    "tailwindcss": "^4.0.14",
+    "ui-mathilde-web": "^0.0.6",
     "yup": "^1.6.1"
   },
   "devDependencies": {
@@ -37,7 +34,8 @@
     "@types/react": "^19.0.1",
     "@types/react-dom": "^19.0.1",
     "autoprefixer": "^10.4.20",
-    "postcss": "^8.4.49",
+    "prettier-plugin-tailwindcss": "^0.6.11",
+    "tailwindcss": "^3.4.17",
     "typescript": "^5.7.2",
     "vite": "^5.4.11",
     "vite-tsconfig-paths": "^5.1.4"
