+++ tailwind.config.ts
@@ -1,25 +1,11 @@
-import type { Config } from "tailwindcss";
-
+ /** @type {import('tailwindcss').Config} */
 export default {
-  content: ["./app/**/{**,.client,.server}/**/*.{js,jsx,ts,tsx}",
-    "../node_modules/flowbite/**/*.js",
-    "../node_modules/flowbite-datepicker/**/*.js"
+  content: [
+    "./index.html",
+    "./app/**/*.{js,jsx,ts,tsx}"
   ],
-  theme: {
-    extend: {
-      fontFamily: {
-        sans: [
-          '"Inter"',
-          "ui-sans-serif",
-          "system-ui",
-          "sans-serif",
-          '"Apple Color Emoji"',
-          '"Segoe UI Emoji"',
-          '"Segoe UI Symbol"',
-          '"Noto Color Emoji"',
-        ],
-      },
-    },
-  },
-  plugins: [require("prettier-plugin-tailwindcss"), require("flowbite/plugin")],
-} satisfies Config;
+   theme: {
+     extend: {},
+   },
+   plugins: [],
+ }
